[test][omdb] Wait for producer registration before listing them (#7183)
Refactors "wait_for_producer" into nexus-test-utils so it can be used by
omdb, outside the nexus integration test suite.

Fixes https://github.com/oxidecomputer/omicron/issues/7182
